Import automatic tile reduction
In September 2016, Johnathan Roatch (@jroatch on GitHub) wrote a
tool that takes a collection of NES 2bpp tiles and finds similar
tiles, as a form of lossy compression.
https://forums.nesdev.com/viewtopic.php?f=21&t=14807

I turned the tool into a library that other Python programs can
call, and I'm about to expose it as an option for savtool.

Also update pilbmp2nes.py with new "value to add to each pixel"
option that was made for another project.